refactor(e2e): move struct and utils from `toolkit/check/mod.rs` to dedicated module
